More util to util
